requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7
pandas==2.1.0
scikit-learn==1.3.0
rapidfuzz==3.2.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
